W5: improve description to the last task and renumber
